+++ src/main/java/com/flint/flint/idcard/domain/IdCard.java
@@ -1,11 +1,10 @@
-package com.flint.flint.member.domain.idcard;
+package com.flint.flint.idcard.domain;
 
 import com.flint.flint.common.BaseTimeEntity;
 import com.flint.flint.member.domain.main.Member;
-import com.flint.flint.member.spec.InterestType;
-import com.flint.flint.member.spec.InterestConverter;
+import com.flint.flint.idcard.spec.InterestType;
+import com.flint.flint.idcard.spec.InterestConverter;
 import jakarta.persistence.*;
-import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Getter;
