+++ src/main/java/com/flint/flint/idcard/repository/IdCardRepository.java
@@ -1,6 +1,6 @@
-package com.flint.flint.member.repository;
+package com.flint.flint.idcard.repository;
 
-import com.flint.flint.member.domain.idcard.IdCard;
+import com.flint.flint.idcard.domain.IdCard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface IdCardRepository extends JpaRepository<IdCard, Long> {
